Add error handler and refactor genres to controller
