+++ src/routes/genre.js
@@ -1,96 +1,13 @@
 const express = require('express')
-const { body, validationResult } = require('express-validator');
-const GenresService = require('../services/genres')
+const { body } = require('express-validator')
 const router = new express.Router()
-const verifyJWT = require('../middleware/verifyJWT')
+const verifyJWT = require('../middleware/verifyJWT')  // TODO: Add to private routes
+const genreController = require('../controllers/genreController')
 
-/**
- *  GET ALL GENRES
- */
-router.get('/genres',verifyJWT, async (req, res) => {
-  try {
-    const genresService = new GenresService()
-    const genres = await genresService.getAll()
-    res.send({genres})
-  } catch(error) {
-    console.log(error)
-    res.status(500).send(error)
-  }
-})
-
-/**
- *  GET GENRE BY ID
- */
-router.get('/genres/:id', async (req, res) => {
-  try {
-    const { params: { id: genreId }} = req;
-    const genresService = new GenresService()
-    const genre = await genresService.getById(genreId)
-    res.send(genre)
-  } catch(error) {
-    console.log(error)
-    res.status(500).send(error)
-  }
-})
-
-/**
- *  ADD GENRE
- */
-router.post('/genres/add', body('name').notEmpty(), async (req, res) => {
-  try {
-    const errors = validationResult(req)
-
-    if (!errors.isEmpty()) {
-      res.status(400).send("Invalid object!")
-      return
-    }
-
-    const genresService = new GenresService()
-    const message = await genresService.add(req.body)
-    res.status(201).send(message)
-  } catch(error) {
-    console.log(error)
-    res.status(500).send(error)
-  }
-})
-
-/**
- *  UPDATE GENRE
- */
-router.put('/genres/edit/:id', body('updatedName').notEmpty(), async (req, res) => {
-  try {
-    const errors = validationResult(req)
-
-    if (!errors.isEmpty()) {
-      res.status(400).send("Invalid object!")
-      return
-    }
-
-    const { params: { id: genreId }} = req;
-    const genresService = new GenresService()
-    const message = await genresService.update(genreId, req.body)
-    res.send(message)
-  } catch(error) {
-    console.log(error)
-    res.status(500).send(error)
-  }
-})
-
-
-
-/**
- *  REMOVE GENRE
- */
-router.delete('/genres/remove/:id', async (req, res) => {
-  try {
-    const { params: { id: genreId }} = req;
-    const genresService = new GenresService()
-    const message = await genresService.remove(genreId)
-    res.send(message)
-  } catch (error) {
-    console.log(error)
-    res.status(500).send(error)
-  }
-})
+router.get('/genres', genreController.getAll)
+router.get('/genres/:id', genreController.getById)
+router.post('/genres/add', body('name').notEmpty(), genreController.add)
+router.put('/genres/edit/:id', body('newName').notEmpty(), genreController.update)
+router.delete('/genres/remove/:id', genreController.remove)
 
 module.exports = router
